// rest-countries-api/src/types/country.ts
export interface Country {
	name: {
		common: string;
		official: string;
		nativeName: {
			[key: string]: {
				official: string;
				common: string;
			};
		};
	};
	tld: string[];
	cca3: string;
	currencies: {
		[key: string]: {
			name: string;
			symbol: string;
		};
	};
	capital: string[];
	region: string;
	subregion: string;
	languages: {
		[key: string]: string;
	};
	borders: string[];
	population: number;
	flags: {
		png: string;
		alt: string;
	};
}